# docs/swagger.yaml
swagger: '2.0'
info:
  description: <Please edit your swagger document to add a description here.>
  version: 1.0.0
  title: category
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  '/category/{id}':
    get:
      tags:
        - 'category/{id}'
      summary: Fetch resource by Id.
      description: ''
      operationId: 'getcategory/{id}'
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: path
          description: 'The id that needs to be fetched. '
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/category'
        '405':
          description: Invalid input
    put:
      tags:
        - 'category/{id}'
      summary: Updates an existing resource.
      description: ''
      operationId: updatecategory
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: path
          description: id that need to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated category object
          required: true
          schema:
            $ref: '#/definitions/category'
      responses:
        '200':
          description: Successfully updated data
        '400':
          description: Invalid input
    delete:
      tags:
        - 'category/{id}'
      summary: Deletes resource by Id.
      description: ''
      operationId: category
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: path
          description: id that need to be deleted
          required: true
          type: string
      responses:
        '200':
          description: Successfully deleted data
        '400':
          description: Invalid input
definitions:
...
  category:
    type: object
    properties:
      name:
        type: string
      createdDate:
        type: string
      modifiedDate:
        type: string
    xml:
      name: category